Added support for subversion
To this plugin, added dependency to the plugin org.jenkins-ci.plugins:subversion
If the repository being scanned is subversion, use svn API to check if the
revisitions match.

One problematic part, I had to use reflection to get the SVNRepository
interface from the SCMFileSystem. Having got this, could call the log()
method to get the commits.

The required methods seem to be private
and/or not implemented. For instance, it seemed like the SCMFileSystem::
changesSince() method would be useful to use, but it is not implemented
for the SubversionSCMFileSystem class.